requirements.txt
appdirs==1.4.3
certifi==2020.6.20
click==7.1.2
distlib==0.3.0
Flask-Cors==3.0.8
gevent==20.6.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
pillow==7.2.0
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
scipy==1.2.1
tensorflow==2.4.1
six==1.14.0
virtualenv==20.0.7
Werkzeug==1.0.1
keras==2.4.3
wincertstore==0.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.9
Flask-gTTS==0.18
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3